Update clients.xml to not use FQDNs
Because we are using a bcfg2 managed /etc/hosts where the short name is
first, the short name is what bcfg2 sees. It then can't match
the hostname to anything in the previous clients.xml so it created
a new basic one.

This patch fixes this issue by replacing FQDNs with the short hostname.