Optimise worker MPM settings
We want to ensure a high level of throughput for most connections,
but limit large writes from the JobWrapper to avoid overloading
BigCouch (it doesn't like large numbers of concurrent writes with
large >~4MB documents).

The trick is to define two separate load balancers composed of the
same workers, but with different parameters. HTTPD doesn't let
you do this (it makes the assumption that a worker will only ever
belong to a single balancer, so they cannot be defined multiple times)
but we can use host names for one balancer, and FQDN for the other,
which allows us to do what we want.

So, balancer bigcouch is for normal use and this is what all CouchDB
ports use as standard. It uses the default maximum for connections
to workers, and prefers bc-37-00, then bc-37-02 then bc-37-01.

Balancer bigcouch-jw is for the jobwrapper only and it is only the
port configured for SSL authentication that uses this balancer. The
connections are limited to 2 per worker, and it prefers bc-37-01,
then bc-37-02, then bc-37-01.

The server configuration in httpd.conf is set up to ensure that
the bigcouch-jw balancer will not exceed the number of concurrent
connections known to create problems with large document writes.
These parameters are:

 * ServerLimit 15
 * ThreadLimit 133
 * MaxClients 1995
 * ThreadsPerChild 133

This allows us to utilise up to 20GB of system RAM, but limits
total number of connections to each bigcouch-jw worker to 30
(15 servers, 2 connections per server).